fetch rates from api && move rates page to page directory &&  fix rates page style
